requirements.txt
Flask==3.0.0
yfinance==0.2.33
pandas==2.1.3
numpy==1.26.2
anthropic==0.8.1
python-dotenv==1.0.0
